fix: reorganizing tests
